cold_bot/requirements.txt
playwright
playwright-stealth
ollama
pyyaml
yagmail
pytest
beautifulsoup4
lxml
black
flake8
httpx
requests
